pyproject.toml
[project]
name = "kalshi"
version = "0.1.0"
description = "Kalshi API client library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kalshi_api"]

